+++ src/components/stable-component.ts
