+++ src/index.ts
@@ -1,11 +1,12 @@
 export { useLucyState } from "./use-lucy-state";
-export { useCombine } from "./combine";
-export { useSelect } from "./select";
+export { useCombine } from "./utils/combine";
+export { useSelect } from "./utils/select";
 export {
   useConvertToLucyState,
   useConvertLucyStateToProperty,
 } from "./convert-into-lucy-state";
-export { StableComponent } from "./stable-component";
-export { StableIteratorComponent } from "./stable-iterator-component";
+
+export { StableComponent } from "./components/stable-component";
+export { StableIteratorComponent } from "./components/stable-iterator-component";
 
 export type { LucyState } from "./types";
